requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SocketIO==5.3.6
Werkzeug==3.0.3
cryptography==42.0.8
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
pytest==8.2.2
pytest-cov==5.0.0
flake8==7.0.0
gunicorn==22.0.0
python-magic==0.4.27
Pillow==10.4.0
boto3==1.34.0
pyotp==2.9.0
qrcode[pil]==7.4.2
PyJWT==2.8.0
simple-websocket==1.0.0
